move exampleData.js to src
